+++ YandexMobileAdsExample/app/src/main/java/com/yandex/ads/sample/tv/instream/player/model/ContentType.kt
@@ -1,0 +1,7 @@
+package com.yandex.ads.sample.tv.instream.player.model
+
+sealed interface ContentType {
+    object Content : ContentType
+    object Ad : ContentType
+    data class Error(val type: ErrorType) : ContentType
+}