+++ YandexMobileAdsExample/app/src/main/java/com/yandex/ads/sample/tv/instream/presentation/InstreamAction.kt
@@ -7,4 +7,5 @@ sealed interface InstreamAction {
     object Pause : InstreamAction
     object SeekBack : InstreamAction
     object SeekForward : InstreamAction
+    object TryAgain : InstreamAction
 }