+++ changelogs/mediation/digitalturbine/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 All notable changes to Digital Turbine Adapter for Yandex Mobile Ads Mediation will be documented in this file.
 
+## Version 8.3.8.0
+
+#### Updated
+* Yandex Mobile Ads SDK version 7.16.1
+* Digitalturbine SDK version 8.3.8
+
 ## Version 8.3.7.2
 
 #### Updated