+++ changelogs/mediation/google/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to AdMob Adapter for Yandex Mobile Ads Mediation will be documented in this file.
 
+## Version 23.6.0.8
+
+#### Updated
+* Yandex Mobile Ads SDK version 7.16.1
+
 ## Version 23.6.0.7
 
 #### Updated